Script to do previous GDEF changes
The glyphs for the Limbu codepoints (U+1921 and U+1939)
are not modified by the script. Font Bakery still reports
GDEF issues for them (they are non-spacing marks in Unicode
but not in the GDEF mark class). The font could be changed
to have them as marks and position them if the warning from
Font Bakery needs to go away.