Improve spacing within some conjuncts
LibreOffice Writer does not seems to use the kern feature by default,
so the dist feature was used instead. Other applications seemed to
enable kern by default, so this change will not affect other applications.

Also increased the height used for clipping, the highest combination
seems to be <consonant, okar, kemphreng>.

Improved test data for testing all of the above.